Move App.js to parent folder
